+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 HEAD
 ---
 * initial settings selector box is big enough to see some fields
+* combo box and text entry now share "staged" column
 
 0.9.0.2
 ---